// src/api/models/enums/DeviceCapabilities.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export enum DeviceCapabilities {
    TEMPERATURE = 'temperature',
    HUMIDITY = 'humidity',
    PRESSURE = 'pressure',
    LIGHT = 'light',
    MOTION = 'motion',
    CO2 = 'co2',
    RELAY = 'relay',
    GPS = 'gps',
    AMMONIA = 'ammonia',
}